requirements.txt
Adafruit-Blinka==8.68.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-bh1750==1.1.17
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-dht==4.0.10
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.85.0
Adafruit-PureIO==1.1.11
attrs==25.4.0
av==16.0.1
binho-host-adapter==0.1.6
blinker==1.9.0
click==8.3.1
colorzero==2.0
Flask==3.1.2
flask-cors==6.0.1
gpiozero==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lgpio==0.2.2.0
libarchive-c==5.3
MarkupSafe==3.0.3
numpy==2.3.5
OpenEXR==3.4.4
picamera2==0.3.33
pidng==4.0.9
piexif==1.1.3
pillow==12.0.0
pyftdi==0.57.1
pyserial==3.5
python-dotenv==1.2.1
python-prctl==1.8.1
pyusb==1.3.1
referencing==0.37.0
rpds-py==0.30.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
simplejpeg==1.9.0
smbus2==0.5.0
sysv-ipc==1.1.0
tqdm==4.67.1
typing_extensions==4.15.0
videodev2==0.0.4
Werkzeug==3.1.3
